refactor(flow): split flow inspection
